hear/pyproject.toml
[project]
name = "sunstone-hear"
version = "0.1.0"
description = "Audio capture and Gemini-based transcription utilities."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv",
    "google-genai",
    "soundcard",
    "soundfile",
    "silero-vad",
    "deepfilternet",
    "torch",
    "librosa",
    "scipy",
    "numpy",
    "websockets",
]

[project.scripts]
gemini-mic = "hear.capture:main"
gemini-transcribe = "hear.transcribe:main"
gemini-live = "hear.live:main"
